+++ src/main/kotlin/com/broeskamp/postident/PostIdentApi.kt
@@ -6,7 +6,7 @@ import com.broeskamp.postident.dto.request.signing.SigningCaseRequest
 import com.broeskamp.postident.dto.response.ident.IdentCaseResponse
 import com.broeskamp.postident.dto.response.signing.SigningCaseResponse
 import com.broeskamp.postident.dto.result.ident.IdentCaseResult
-import com.broeskamp.postident.dto.result.sign.SigningCaseResult
+import com.broeskamp.postident.dto.result.signing.SigningCaseResult
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.databind.SerializationFeature
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
