+++ src/main/kotlin/com/broeskamp/postident/dto/result/signing/AddressResult.kt
@@ -1,27 +1,19 @@
-package com.broeskamp.postident.dto.result.sign
+package com.broeskamp.postident.dto.result.signing
 
 import com.thinkinglogic.builder.annotation.Builder
 
 @Builder
 data class AddressResult(
-    /**
-     * Max value length: 55
-     */
+    /** Max value length: 55 */
     val streetAddress: ResultValue?,
 
-    /**
-     * Max value length: 55
-     */
+    /** Max value length: 55 */
     val appendix: ResultValue?,
 
-    /**
-     * Max value length: 11
-     */
+    /** Max value length: 11 */
     val postalCode: ResultValue?,
 
-    /**
-     * Max value length: 55
-     */
+    /** Max value length: 55 */
     val city: ResultValue?,
 
     /**
