+++ src/main/kotlin/com/broeskamp/postident/dto/result/signing/ResultValue.kt
@@ -1,4 +1,4 @@
-package com.broeskamp.postident.dto.result.sign
+package com.broeskamp.postident.dto.result.signing
 
 import com.fasterxml.jackson.annotation.JsonProperty
 import com.thinkinglogic.builder.annotation.Builder
@@ -10,8 +10,10 @@ data class ResultValue(
 ) {
     /**
      * Possible outcomes:
-     *
-     * - unchecked: value has not been verified during the identification process. This happens for example if the identification document doesn't comprise a street address. In this case the value is returned which was passed in on case creation.
+     * - unchecked: value has not been verified during the identification
+     *   process. This happens for example if the identification document
+     *   doesn't comprise a street address. In this case the value is returned
+     *   which was passed in on case creation.
      * - new: value was not provided by client
      * - match: value as provided by client
      * - change: value was modified during identification process
